Refactor domain folders to split packages from modules
