+++ src/domain/packages/entities/Errors.ts
