+++ src/domain/packages/entities/Store.ts
