+++ src/domain/packages/entities/StorePermissions.ts
