+++ src/domain/packages/usecases/ValidateStoreUseCase.ts
