+++ src/presentation/CompositionRoot.ts
@@ -25,19 +25,19 @@ import { ImportMetadataUseCase } from "../domain/metadata/usecases/ImportMetadat
 import { ListAllMetadataUseCase } from "../domain/metadata/usecases/ListAllMetadataUseCase";
 import { ListMetadataUseCase } from "../domain/metadata/usecases/ListMetadataUseCase";
 import { MetadataSyncUseCase } from "../domain/metadata/usecases/MetadataSyncUseCase";
-import { CreatePackageUseCase } from "../domain/modules/usecases/CreatePackageUseCase";
 import { DeleteModuleUseCase } from "../domain/modules/usecases/DeleteModuleUseCase";
-import { DeletePackageUseCase } from "../domain/modules/usecases/DeletePackageUseCase";
 import { DownloadModuleSnapshotUseCase } from "../domain/modules/usecases/DownloadModuleSnapshotUseCase";
-import { DownloadPackageUseCase } from "../domain/modules/usecases/DownloadPackageUseCase";
 import { GetModuleUseCase } from "../domain/modules/usecases/GetModuleUseCase";
-import { GetPackageUseCase } from "../domain/modules/usecases/GetPackageUseCase";
-import { GetStoreUseCase } from "../domain/modules/usecases/GetStoreUseCase";
 import { ListModulesUseCase } from "../domain/modules/usecases/ListModulesUseCase";
-import { ListPackagesUseCase } from "../domain/modules/usecases/ListPackagesUseCase";
 import { SaveModuleUseCase } from "../domain/modules/usecases/SaveModuleUseCase";
-import { SaveStoreUseCase } from "../domain/modules/usecases/SaveStoreUseCase";
-import { ValidateStoreUseCase } from "../domain/modules/usecases/ValidateStoreUseCase";
+import { CreatePackageUseCase } from "../domain/packages/usecases/CreatePackageUseCase";
+import { DeletePackageUseCase } from "../domain/packages/usecases/DeletePackageUseCase";
+import { DownloadPackageUseCase } from "../domain/packages/usecases/DownloadPackageUseCase";
+import { GetPackageUseCase } from "../domain/packages/usecases/GetPackageUseCase";
+import { GetStoreUseCase } from "../domain/packages/usecases/GetStoreUseCase";
+import { ListPackagesUseCase } from "../domain/packages/usecases/ListPackagesUseCase";
+import { SaveStoreUseCase } from "../domain/packages/usecases/SaveStoreUseCase";
+import { ValidateStoreUseCase } from "../domain/packages/usecases/ValidateStoreUseCase";
 import { Repositories } from "../domain/Repositories";
 import { DownloadFileUseCase } from "../domain/storage/usecases/DownloadFileUseCase";
 import { D2 } from "../types/d2";
