+++ src/presentation/webapp/components/module-wizard/common/MetadataSelectionStep.tsx
@@ -1,7 +1,7 @@
 import { useSnackbar } from "d2-ui-components";
 import _ from "lodash";
 import React, { useCallback } from "react";
-import { MetadataModule } from "../../../../../domain/modules/entities/modules/MetadataModule";
+import { MetadataModule } from "../../../../../domain/modules/entities/MetadataModule";
 import i18n from "../../../../../locales";
 import { DashboardModel, DataSetModel, ProgramModel } from "../../../../../models/dhis/metadata";
 import MetadataTable from "../../metadata-table/MetadataTable";
